+++ engines/avp/assets/run-avp.sh
@@ -13,8 +13,17 @@ create_relative_symlink () {
         ln -rsf "$target" "$symlink"
 }
 
+create_relative_symlink_fmvs () {
+        local -r target="$(basename "$1")"
+        local -r baselower="$(lowercase "$target")"
+        if [ "$target" != "$baselower" ]; then
+			local -r symlink="FMVs/$(lowercase "$target")"
+			ln -rsf "$1" "$symlink"
+        fi
+}
+
 if [ ! -f ready ]; then
-	find {avp_huds,avp_rifs,fastfile,FMVs,Text} -type f  | while read -r file_name ; do
+	find {avp_huds,avp_rifs,fastfile,Text} -type f  | while read -r file_name ; do
         	create_relative_symlink "$file_name"
 	done
 
@@ -23,6 +32,18 @@ if [ ! -f ready ]; then
 	ln -rsf "CD Tracks.txt" dataconv/"cd tracks.txt"
 
 	touch ready
+fi
+
+if [ ! -f fmv_ready ]; then
+	find FMVs -type f  | while read -r file_name ; do
+		create_relative_symlink "$file_name"
+	done
+
+	find FMVs -type f  | while read -r file_name ; do
+		create_relative_symlink_fmvs "$file_name"
+	done
+
+	touch fmv_ready
 fi
 
 AVP_DATA=dataconv LD_LIBRARY_PATH=./lib ./avp --fullscreen