+++ engines/daggerfall-unity/assets/run-daggerfall-unity.sh
@@ -10,14 +10,6 @@ if [ ! -d DFUPDATED ]; then
     mkdir -p DFUPDATED/DAGGER
     LD_PRELOAD="" cp -rfv ./DF/DAGGER ./DFUPDATED
     LD_PRELOAD="" ln -rsf DF/DFCD/DAGGER/ARENA2/*.VID DFUPDATED/DAGGER/ARENA2
-    cp -rfv "$ORIGINALPWD/dag213.exe" ./DFUPDATED/DAGGER/dag213.exe
-    cp -rfv "$ORIGINALPWD/dfpatch.bat" ./DFUPDATED/dfpatch.bat #built based on instructions from https://www.dfworkshop.net/using-steam-release-of-daggerfall-with-daggerfall-unity/
-
-    "$STEAM_ZENITY" --info --title="Daggerfall Patch Setup" --text="dosbox-staging is used to update the game to the latest official patch. Follow the on-screen prompts and confirm each one. Once it is complete, dosbox-staging will close and the game will launch."
-
-    pushd DFUPDATED
-        LD_LIBRARY_PATH="$ORIGINALPWD/dosbox-staging/lib:$LD_LIBRARY_PATH" "$ORIGINALPWD/dosbox-staging/dosbox" ./dfpatch.bat -exit
-    popd
 fi
 
 if [[ ! -z "${DEPPATH_1070560}" ]]; then