+++ engines/dhewm3/build.sh
@@ -3,7 +3,7 @@
 # CLONE PHASE
 git clone https://github.com/dhewm/dhewm3.git source
 pushd source
-git checkout 441c26e
+git checkout 2e71b99
 popd
 
 # BUILD PHASE
@@ -21,3 +21,4 @@ popd
 # COPY PHASE
 cp -rfv tmp/bin/* "$diststart/common/dist/"
 cp -rfv tmp/lib/dhewm3/* "$diststart/common/dist/"
+cp -rfv assets/* "$diststart/common/dist/"