+++ engines/openxcom-oxce/assets/run-ufo.sh
@@ -7,4 +7,17 @@ cd "$DIR"
 rm -rf ./UFO
 ln -rsf "$originalpwd/XCOM" ./UFO
 
+if ! [[ -z "${LUX_STEAM_DECK}" ]]; then
+    if [ ! -f ~/.config/openxcom/options.cfg ]; then
+        if [ ! -d ~/.config/openxcom ]; then
+            mkdir -p ~/.config/openxcom
+        fi
+        echo "No options.cfg file detected, so creating"
+        echo -e "options:" >> ~/.config/openxcom/options.cfg
+        echo -e "  displayHeight: 800" >> ~/.config/openxcom/options.cfg
+        echo -e "  displayWidth: 1280" >> ~/.config/openxcom/options.cfg
+        echo -e "  fullscreen: false" >> ~/.config/openxcom/options.cfg
+    fi
+fi
+
 LD_LIBRARY_PATH=./lib:$LD_LIBRARY_PATH ./openxcom --data ./