+++ engines/serious-engine/assets/run-serious-engine.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cp -rfv /lib/x86_64-linux-gnu/libvorbisfile.so.3.3.7 ./Bin/libvorbisfile.so
+
+./Bin/SeriousSam "$@"