+++ engines/ut2004/build.sh
@@ -72,9 +72,7 @@ fi
 popd
 
 # COPY PHASE
-cp -rfv "assets/run-ut2004.sh" "$diststart/13230/dist/"
-cp -rfv "assets/setup-ut2004.sh" "$diststart/13230/dist/"
-cp -rfv "assets/uninstall-ut2004.sh" "$diststart/13230/dist/"
+cp -rfv assets/* "$diststart/13230/dist/"
 
 cp -rfv "openal/build/libopenal.so.1.20.1" "$diststart/13230/dist/openal.so"
 cp -rfv "$pfx/lib/libSDL-1.2.so.1.2.52" "$diststart/13230/dist/libSDL-1.2.so.0"