+++ engines/warzone2100/assets/run-warzone2100.sh
@@ -6,4 +6,16 @@ cd "$DIR"
 
 ln -rsf ../data/sequences.wz ./data/sequences.wz
 
+if ! [[ -z "${LUX_STEAM_DECK}" ]]; then
+    if [ ! -f ~/.local/share/warzone2100/config ]; then
+        if [ ! -d ~/.local/share/warzone2100 ]; then
+            mkdir -p ~/.local/share/warzone2100
+        fi
+        echo -e "[general]" >> ~/.local/share/warzone2100/config
+        echo -e "fullscreen=1" >> ~/.local/share/warzone2100/config
+        echo -e "width=1280" >> ~/.local/share/warzone2100/config
+        echo -e "height=800" >> ~/.local/share/warzone2100/config
+    fi
+fi
+
 LD_LIBRARY_PATH=lib:$LD_LIBRARY_PATH bin/warzone2100